Build and test collections on Kotlin/JS
