+++ collection/collection/src/commonTest/kotlin/androidx/collection/ArraySetTest.kt
@@ -15,100 +15,14 @@
  */
 package androidx.collection
 
-import androidx.collection.internal.Lock
-import androidx.collection.internal.synchronized
-import kotlin.coroutines.CoroutineContext
-import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.isActive
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.test.runTest
 
 internal class ArraySetTest {
     private val set = ArraySet<String>()
-
-    /**
-     * Attempt to generate a ConcurrentModificationException in ArraySet.
-     *
-     * ArraySet is explicitly documented to be non-thread-safe, yet it's easy to accidentally screw
-     * this up; ArraySet should (in the spirit of the core Java collection types) make an effort to
-     * catch this and throw ConcurrentModificationException instead of crashing somewhere in its
-     * internals.
-     */
-    @Test
-    fun testConcurrentModificationException() = runTest {
-        var error: Throwable? = null
-        val nThreads = 20
-        var nActiveThreads = 0
-        val lock = Lock()
-        val context: CoroutineContext = Dispatchers.Default
-        val scope = CoroutineScope(context)
-
-        repeat(nThreads) {
-            scope.launch {
-                lock.synchronized { nActiveThreads++ }
-
-                while (isActive) {
-                    val add = Random.nextBoolean()
-                    try {
-                        if (add) {
-                            set.add(Random.nextLong().toString())
-                        } else {
-                            set.clear()
-                        }
-                    } catch (e: IndexOutOfBoundsException) {
-                        if (!add) {
-                            error = e
-                            throw e
-                        } // Expected exception otherwise
-                    } catch (e: ClassCastException) {
-                        error = e
-                        throw e
-                    } catch (ignored: ConcurrentModificationException) {
-                        // Expected exception
-                    }
-                }
-            }
-        }
-
-        // Wait until all worker threads are started
-        for (i in 0 until 100) {
-            if (lock.synchronized { nActiveThreads == nThreads }) {
-                break
-            } else {
-                delay(timeMillis = 10L)
-            }
-        }
-
-        // Allow the worker threads to run concurrently for some time
-        delay(timeMillis = 100L)
-
-        scope.cancel()
-
-        error?.also { throw it }
-    }
-
-    /** Check to make sure the same operations behave as expected in a single thread. */
-    @Test
-    fun testNonConcurrentAccesses() {
-        repeat(100_000) { i ->
-            set.add("key $i")
-            if (i % 200 == 0) {
-                print(".")
-            }
-            if (i % 500 == 0) {
-                set.clear()
-            }
-        }
-    }
 
     @Test
     fun testCanNotIteratePastEnd() {
