+++ collection/collection/src/commonTest/kotlin/androidx/collection/FloatIntMapTest.kt
@@ -619,7 +619,7 @@ class FloatIntMapTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableFloatIntMap()
         map[1f] = 1
