+++ collection/collection/src/commonTest/kotlin/androidx/collection/FloatObjectMapTest.kt
@@ -656,7 +656,7 @@ class FloatObjectMapTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableFloatObjectMap<String?>()
         map[1f] = "World"
         map[2f] = null
