+++ collection/collection/src/commonTest/kotlin/androidx/collection/IntIntMapTest.kt
@@ -619,7 +619,7 @@ class IntIntMapTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableIntIntMap()
         map[1] = 1
