+++ collection/collection/src/commonTest/kotlin/androidx/collection/IntObjectMapTest.kt
@@ -656,7 +656,7 @@ class IntObjectMapTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableIntObjectMap<String?>()
         map[1] = "World"
         map[2] = null
