+++ collection/collection/src/commonTest/kotlin/androidx/collection/LongIntMapTest.kt
@@ -619,7 +619,7 @@ class LongIntMapTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableLongIntMap()
         map[1L] = 1
