+++ collection/collection/src/commonTest/kotlin/androidx/collection/LongObjectMapTest.kt
@@ -656,7 +656,7 @@ class LongObjectMapTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableLongObjectMap<String?>()
         map[1L] = "World"
         map[2L] = null
