+++ collection/collection/src/commonTest/kotlin/androidx/collection/LongSetTest.kt
@@ -367,7 +367,7 @@ class LongSetTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val set = MutableLongSet()
         set += 1L
         set += 5L
