+++ collection/collection/src/commonTest/kotlin/androidx/collection/ObjectFloatMapTest.kt
@@ -635,7 +635,7 @@ class ObjectFloatTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableObjectFloatMap<String?>()
         map["Hello"] = 1f
         map[null] = 2f
