+++ collection/collection/src/commonTest/kotlin/androidx/collection/ObjectIntMapTest.kt
@@ -635,7 +635,7 @@ class ObjectIntTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val map = MutableObjectIntMap<String?>()
         map["Hello"] = 1
         map[null] = 2
