+++ collection/collection/src/commonTest/kotlin/androidx/collection/ScatterSetTest.kt
@@ -560,7 +560,7 @@ class ScatterSetTest {
     }
 
     @Test
-    fun equals() {
+    fun equalsTest() {
         val set = MutableScatterSet<String?>()
         set += "Hello"
         set += null
